// type.d.ts
export interface EVData {
    'Make': string;
    'Model': string;
    'Model Year': string;
    'City': string;
    'County': string;
    'State': string;
    'Electric Range': string;
    'Electric Vehicle Type': string;
    'Clean Alternative Fuel Vehicle (CAFV) Eligibility': string;
    'DOL Vehicle ID':string;
    'CAFVEligibility':string;
  
  }
  